Summarize the Result of an SVD Cross-Validation

Summarize the result of cv.svd.gabriel or cv.svd.wold.

# S3 method for cvsvd
summary(object, ...)

Arguments

object

the result of a cv.svd.gabriel or cv.svd.wold computation.

...

additional arguments to summary.

Value

nfolds

the number of cross-validation folds

maxrank

the maximum rank for which prediction error is estimated.

msep.mean

the average mean square error of prediction (MSEP) across all folds for ranks 0, 1, ..., maxrank.

msep.se

the standard errors of the msep.mean estimates.

rank.best

the rank with the minimum msep.mean value.

rank.1se

the smallest rank within one standard error of the minimum msep.mean value.

Details

Print a table of the estimated prediction errors and the standard errors of the estimate. Put an asterisk (*) next to the minimum and a plus (+) next to the "one standard error rule" choice.

See also

Author

Patrick O. Perry